requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Werkzeug==2.3.6
gunicorn==20.1.0
cryptography
requests
dnspython
pyotp
qrcode[pil]
psycopg2-binary
